+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## Change Log
 
 ### [Unreleased][unreleased]
+#### Added
+- Multi write connections support
 
 ### [1.5.0] - 2025-04-01
 #### Added